+++ fabric/src/main/java/com/devnemo/nemos/campfires/FabricNemosCampfires.java
@@ -11,24 +11,35 @@ public class FabricNemosCampfires implements ModInitializer {
         NemosCampfiresCommon.init();
 
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.ACACIA_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.BIRCH_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CHERRY_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.PALE_OAK_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CRIMSON_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.DARK_OAK_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.JUNGLE_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.MANGROVE_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.SPRUCE_CAMPFIRE.get());
-        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.WARPED_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get());
         BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.WARPED_SOUL_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.ACACIA_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.BIRCH_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CHERRY_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.PALE_OAK_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.CRIMSON_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.DARK_OAK_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.JUNGLE_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.MANGROVE_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.OAK_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.SPRUCE_COPPER_CAMPFIRE.get());
+        BlockEntityType.CAMPFIRE.addSupportedBlock(NemosCampfiresBlocks.WARPED_COPPER_CAMPFIRE.get());
     }
 }