+++ fabric/src/main/java/com/devnemo/nemos/campfires/datagen/lang/EnglishLanguageProvider.java
@@ -4,6 +4,7 @@ import com.devnemo.nemos.campfires.world.level.block.NemosCampfiresBlocks;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricLanguageProvider;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.world.level.block.Blocks;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -18,25 +19,38 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
     @Override
     public void generateTranslations(HolderLookup.Provider registryLookup, TranslationBuilder translationBuilder) {
         translationBuilder.add(NemosCampfiresBlocks.ACACIA_CAMPFIRE.get(), "Acacia Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get(), "Acacia Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.BIRCH_CAMPFIRE.get(), "Birch Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get(), "Birch Soul Campfire");
+        translationBuilder.add(Blocks.CAMPFIRE, "Oak Campfire");
         translationBuilder.add(NemosCampfiresBlocks.CHERRY_CAMPFIRE.get(), "Cherry Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get(), "Cherry Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.PALE_OAK_CAMPFIRE.get(), "Pale Oak Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get(), "Pale Oak Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.CRIMSON_CAMPFIRE.get(), "Crimson Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get(), "Crimson Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.DARK_OAK_CAMPFIRE.get(), "Dark Oak Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get(), "Dark Oak Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.JUNGLE_CAMPFIRE.get(), "Jungle Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get(), "Jungle Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.MANGROVE_CAMPFIRE.get(), "Mangrove Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get(), "Mangrove Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.SPRUCE_CAMPFIRE.get(), "Spruce Campfire");
-        translationBuilder.add(NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get(), "Spruce Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.WARPED_CAMPFIRE.get(), "Warped Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get(), "Acacia Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get(), "Birch Soul Campfire");
+        translationBuilder.add(Blocks.SOUL_CAMPFIRE, "Oak Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get(), "Cherry Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get(), "Pale Oak Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get(), "Crimson Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get(), "Dark Oak Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get(), "Jungle Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get(), "Mangrove Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get(), "Spruce Soul Campfire");
         translationBuilder.add(NemosCampfiresBlocks.WARPED_SOUL_CAMPFIRE.get(), "Warped Soul Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.ACACIA_COPPER_CAMPFIRE.get(), "Acacia Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.BIRCH_COPPER_CAMPFIRE.get(), "Birch Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.CHERRY_COPPER_CAMPFIRE.get(), "Cherry Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.PALE_OAK_COPPER_CAMPFIRE.get(), "Pale Oak Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.CRIMSON_COPPER_CAMPFIRE.get(), "Crimson Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.DARK_OAK_COPPER_CAMPFIRE.get(), "Dark Oak Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.JUNGLE_COPPER_CAMPFIRE.get(), "Jungle Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.MANGROVE_COPPER_CAMPFIRE.get(), "Mangrove Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.OAK_COPPER_CAMPFIRE.get(), "Oak Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.SPRUCE_COPPER_CAMPFIRE.get(), "Spruce Copper Campfire");
+        translationBuilder.add(NemosCampfiresBlocks.WARPED_COPPER_CAMPFIRE.get(), "Warped Copper Campfire");
         translationBuilder.add("itemGroup." + MOD_ID, "Nemo's Campfires");
     }
 }