+++ neoforge/src/main/java/com/devnemo/nemos/campfires/event/EventHandler.java
@@ -14,24 +14,35 @@ public class EventHandler {
     @SubscribeEvent
     public static void onBlockEntityTypeAddBlockEvent(BlockEntityTypeAddBlocksEvent event) {
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.ACACIA_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.BIRCH_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CHERRY_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.PALE_OAK_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CRIMSON_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.DARK_OAK_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.JUNGLE_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.MANGROVE_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.SPRUCE_CAMPFIRE.get());
-        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.WARPED_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.ACACIA_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.BIRCH_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CHERRY_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.PALE_OAK_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CRIMSON_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.DARK_OAK_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.JUNGLE_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.MANGROVE_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.SPRUCE_SOUL_CAMPFIRE.get());
         event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.WARPED_SOUL_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.ACACIA_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.BIRCH_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CHERRY_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.PALE_OAK_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.CRIMSON_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.DARK_OAK_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.JUNGLE_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.MANGROVE_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.OAK_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.SPRUCE_COPPER_CAMPFIRE.get());
+        event.modify(BlockEntityType.CAMPFIRE, NemosCampfiresBlocks.WARPED_COPPER_CAMPFIRE.get());
     }
 }