+++ Changelog.md
@@ -1,4 +1,8 @@
-# Changelog v2.0
+# Changelog v2.1
 
-## Breaking Changes
-- Updated modId to nemos_campfires instead of nemos-campfires
+## Additions
+- Added Copper Campfires
+
+## Changes
+- Renamed Campfire to Oak Campfire
+- Renamed Soul Campfire to Oak Campfire